requirements.txt
# Board Game Tracker - Python Dependencies
# Install with: pip install -r requirements.txt

# Web UI
streamlit>=1.32.0

# Data analysis
pandas>=2.0.0
numpy>=1.26.0

# ML / Recommendation
scikit-learn>=1.4.0

# Charts
plotly>=5.20.0

# Database
mysql-connector-python>=8.3.0

# Utilities
python-dotenv>=1.0.0      # for storing DB credentials in .env file
sshtunnel>=0.4.0          # for automatic SSH tunnel to school MySQL server
